Updated formatting for names.cfg. Removed the randName includes from Distro and all isntances of use replaced with NPC_SetName function in npcInit.inc from brainAI. npcdesc.cfg updated for <random> name templates, having NameTemplate entries for new names.cfg.
ToDo: Update BrainAI documentation to reflect that names.cfg groups for human graphic (male and female), need to have _Male and _Female on them respectively. The random name code in brain AI now checks graphic for male and female. It ADDS the _Male and _Female to the end of NameTemplate entry. So in the npctemplates, do NOT include _Male and _Female on the end if they are human npcs.