Update for $Id$ property.
New moongates package. Upgraded it for autoload.cfg, and also rewrote the src. Now uses a formatted config for if a realm should be generated for moongates and added to the menu on the gump.

Still need to make the gump use this new system. Also, need to finish testing out the moongates pkg. Testing was limited for me, so recommend more.