Runebook and spell stuff.
Spellstarter modified to freeze during cast/delay. Target won't check if they move. If scroll is passed as INT 1, it overrides reg and scroll and casts without scroll consumption or requiring reagents. Added for runebooks, wands, and so on.

Gate/Mark/Recall added with compatibility to new runebook system.

Still need to make up all the rune colors for the config.

Added ability to check if austin wears little girl's undies.

Added a timebomb for every time Austin mentions polygons and UO in the same sentece.